test: add mock cgo backend
